Remove image server
